requirements.txt
# Основные зависимости
python-dotenv==1.0.0
python-multipart==0.0.9

# Telegram Bot
aiogram~=3.3.0
aiohttp==3.9.1   # ✅ Совместимая с aiogram

# База данных
asyncpg==0.29.0
sqlalchemy[asyncio]~=2.0.44
alembic==1.13.1              # ✅ Последняя версия

# API
fastapi==0.109.2  # ✅ Последняя версия
uvicorn[standard]==0.27.1  # ✅ Последняя версия
pydantic==2.6.1            # ✅ Pydantic 2.x (совместим с последним aiogram)
pydantic-core==2.16.2      # ✅ Соответствующая версия

# Утилиты
pandas==2.2.1              # ✅ Последняя версия
python-dateutil==2.9.0.post0
pytz==2024.1

# Для разработки
pytest==8.0.0              # ✅ Последняя версия
pytest-asyncio==0.23.5
black==24.2.0
isort==5.13.2
PyPDF2~=3.0.1
pytesseract~=0.3.13
pillow~=12.0.0
python-docx~=1.2.0